Add a safe-guard to the test-image setup script
The setup can do any number of dangerous things like nuking parts of
your system, make sure nobody accidentally runs it by placing a flag
file from make_install and checking on setup.sh.

+++ tests/setup.sh
@@ -3,6 +3,12 @@
 # Setup the environment in the test-image.
 # Runs INSIDE the image, with our newly built rpm in the path.
 
+# Make sure nobody runs this accidentally outside the environment
+if [ ! -f /.rpmtestsuite ]; then
+    echo "Not inside rpm test-suite image"
+    exit 99
+fi
+
 mkdir -p /build
 ln -sf ../data/SOURCES /build/
